Rename VERSION -> WORR_VERSION and CI/tool fixes
Rename repository version metadata from VERSION to WORR_VERSION and update version.py and docs accordingly. Replace non-portable std::sinf calls with std::sin across multiple sgame translation units for cross-compiler portability. Improve MSI packaging tooling: build_msi.ps1 now accepts ProductVersion (Version alias), introduces Invoke-NativeCommand to hard-fail on native tool errors, creates output directory earlier, and reliably cleans temporary staging; worr.wxs removes redundant define lines and adds Product Language="1033". Add a recovery implementation log (nightly-ci-cross-platform-recovery-2026-03-16.md) and update related documentation to reflect these changes.

+++ src/game/sgame/gameplay/g_func.cpp
@@ -2516,8 +2516,8 @@ static void BobThink(gentity_t* self) {
 	const int frame = self->bobFrame % cycleTime;
 	const int nextFrame = (self->bobFrame + 1) % cycleTime;
 
-	const float phase0 = std::sinf(2.0f * M_PI * (frame / static_cast<float>(cycleTime)));
-	const float phase1 = std::sinf(2.0f * M_PI * (nextFrame / static_cast<float>(cycleTime)));
+	const float phase0 = std::sin(2.0f * M_PI * (frame / static_cast<float>(cycleTime)));
+	const float phase1 = std::sin(2.0f * M_PI * (nextFrame / static_cast<float>(cycleTime)));
 
 	const float delta = (self->bob / 2.0f) * (phase1 - phase0);
 	self->velocity[_Z] = delta / FRAME_TIME_MS.milliseconds();
@@ -4201,7 +4201,7 @@ static THINK(func_bobbing_think) (gentity_t* ent) -> void {
 	const float phase_offset = ent->phase * cycle;
 	const float frac = fmodf((level.time.milliseconds() * 0.001f + phase_offset), cycle) / cycle;
 	const float angle = frac * M_TWOPI;
-	const float bob = std::sinf(angle) * ent->height;
+	const float bob = std::sin(angle) * ent->height;
 
 	Vector3 delta = { 0, 0, 0 };
 
@@ -4288,7 +4288,7 @@ static THINK(func_pendulum_think) (gentity_t* ent) -> void {
 	const float cycle = ent->speed > 0.0f ? ent->speed : 30.0f; // temporary use for swing angle
 	const float duration = ent->wait; // seconds
 	const float frac = fmodf(level.time.milliseconds() + ent->phase * duration, duration) / duration;
-	const float angle = std::sinf(frac * M_TWOPI) * cycle;
+	const float angle = std::sin(frac * M_TWOPI) * cycle;
 
 	ent->s.angles = ent->pos1; // reset to base position
 	ent->s.angles[ROLL] += angle;
